==> src/00_create_mappers.ipynb <==
# Create Mapper Datasets
## 00_create_mappers

| Date | User | Change Type | Remarks |  
| ---- | ---- | ----------- | ------- |
| 02/10/2025   | Adrienne | Created | Created mapper datasets | 
| 07/10/2025 | Adrienne | Created | Created preventative care mapper |
| 13/10/2025 | Martin | Updated | Code cleanup. Added docstrings to functions |

## Content

* [Introduction](#introduction)
* [Load Data](#load-data)
* [Create Mapper Functions](#create-mapper-functions)
* [Create Maps](#create-maps)
* [Save Datasets](#save-datasets)

## Introduction

The mapper combines the various HCPCS encodings (compiled by Adrienne) in a single dataframe. It expands the coding range from `code_range_start` to `code_range_end` and maps it to the corresponding description.

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
from sklearn.preprocessing import LabelEncoder

## Load Data

path = "../data/raw"
hcpcs_1_map = pd.read_excel(f"{path}/HCPCS Code Ranges.xlsx", sheet_name='HCPCS Level I', dtype=str)
hcpcs_2_map = pd.read_excel(f"{path}/HCPCS Code Ranges.xlsx", sheet_name='HCPCS Level II', dtype=str)
cpt_map = pd.read_excel(f"{path}/HCPCS Code Ranges.xlsx", sheet_name='CPT Category II', dtype=str)
preventative_map = pd.read_excel(f"{path}/Preventative Care Codes.xlsx", sheet_name='Preventative Care', dtype=str)

## Create Mapper Functions

Functions to unify and consolidate HCPCS codes into a single mapper

def extend_values(val: int, maxlen: int):
  """Extends the length of values by prepending "0"s to the front based on the max length

  Args:
      val (int): Values to extend
      maxlen (int): Maximum length of value

  Returns:
      str: Extended value prepended with 0s
  """
  if len(str(val)) < maxlen:
    return (maxlen - len(str(val))) * "0" + str(val)
  else:
    return str(val)

def create_mapper(raw: pd.DataFrame, dset: str):
  """Extracts the codes, category and description of HCPCS values to a single dataframe that
  to consolidate and create a mapper for all codes

  Args:
      raw (pd.DataFrame): Raw HCPCS dataset
      dset (str): Type of dataset (e.g hcpcs_1)

  Returns:
      pd.DataFrame: Output mapper dataframe
  """
  out = {
    'code': [],
    'category': [],
    'description': []
  }
  for idx, row in raw.iterrows():
    if dset == 'hcpcs_1':
      row['code_range_start_int'] = int(row['code_range_start'])
      row['code_range_end_int'] = int(row['code_range_end'])
      
      value = np.arange(row['code_range_start_int'], row['code_range_end_int']+1)
      value = list(map(lambda x: extend_values(x, 5), value))
      out['code'].append(value)
      out['category'].append(len(value) * ["HCPCS_level_1"])
      out['description'].append(len(value) * [row['range_description']])
    
    elif row['code_range_start'][0].isalpha():
      if pd.isna(row['code_range_end']):
        out['code'].append([row['code_range_start']])
        out['category'].append(["HCPCS_level_2"])
        out['description'].append([row['range_description']])
      else:
        try:
          start_letter = row['code_range_start'][0]
          start_num, end_num = int(row['code_range_start'][1:]), int(row['code_range_end'][1:])
          value = np.arange(start_num, end_num+1)
          value = list(map(lambda x: extend_values(x, 4), value))
          value = [f"{start_letter}{i}" for i in value]
          out['code'].append(value)
          out['category'].append(len(value) * ["HCPCS_level_2"])
          out['description'].append(len(value) * [row['range_description']])
        except:
          pass
    
    elif row['code_range_start'][-1].isalpha():
      try:
        end_letter = row['code_range_start'][-1]
        start_num, end_num = int(row['code_range_start'][:-1]), int(row['code_range_end'][:-1])
        value = np.arange(start_num, end_num+1)
        value = list(map(lambda x: extend_values(x, 4), value))
        value = [f"{i}{end_letter}" for i in value]
        out['code'].append(value)
        out['category'].append(len(value) * ["CPT"])
        out['description'].append(len(value) * [row['range_description']])
      except:
        pass

  out = {k: [i for sublist in v for i in sublist] for k, v in out.items()}
  out = pd.DataFrame(out)
  
  return out

## Create Maps

Run the mapper function against the HCPCS and CPT datasets

hcpcs_1_map = create_mapper(hcpcs_1_map, 'hcpcs_1')
hcpcs_2_map = create_mapper(hcpcs_2_map, '')
cpt_map = create_mapper(cpt_map, '')

combined_mapper = pd.concat([hcpcs_1_map, hcpcs_2_map, cpt_map], axis=0)
combined_mapper.head()

### Save Datasets

mapper_path = "../data/mappers"

hcpcs_1_map.to_pickle(f"{mapper_path}/hcpcs_1_map.pkl")
hcpcs_2_map.to_pickle(f"{mapper_path}/hcpcs_2_map.pkl")
cpt_map.to_pickle(f"{mapper_path}/cpt_map.pkl")
combined_mapper.to_pickle(f"{mapper_path}/combined_mapper.pkl")
preventative_map.to_pickle(f"{mapper_path}/preventative_mapper.pkl")
