==> src/00_create_secondary_datasets.ipynb <==
# Creating Secondary Datasets
## 00_create_secondary_datasets

| Date | User | Change Type | Remarks |  
| ---- | ---- | ----------- | ------- |
| 28/09/2025 | Adrienne | Created | Write code to create secondary datasets | 
| 02/10/2025 | Adrienne | Updated | Moved code for code ranges to 00_create_mappers |
| 13/10/2025 | Martin | Updated | Code cleanup |

## Content

* [Introduction](#introduction)
* [Create DRG Dataset](#create-drg-dataset)
* [Create HCPCS Dataset](#create-hcpcs-Dataset)
* [Create ICD-10 Dataset](#create-icd-10-dataset)


## Introduction

The secondary datasets map DRG, HCPCS and ICD-10 codes to the their respective decisions. No additional logic was added, this only serves to unify the formatting

import pandas as pd
import numpy as np
import re

## Create DRG Dataset

# Path to raw excel files
path = "../data/raw"
drg_df = pd.read_excel(f"{path}/DRG_MDC_Code_List.xlsx", dtype=str)

drg_df.tail()

## Create HCPCS Dataset

hcpcs_df = pd.read_excel(f"{path}/HCPCS codes updated 2025.xlsx", dtype=str)

hcpcs_df.tail()

## Create ICD-10 Dataset

icd10_df = pd.read_excel(f"{path}/section111validicd10-jan2025_0.xlsx", dtype=str)

icd10_df.head()

## Save Datasets

mapper_path = "../data/mappers"

drg_df.to_pickle(f"{mapper_path}/drg.pkl")
hcpcs_df.to_pickle(f"{mapper_path}/hcpcs.pkl")
icd10_df.to_pickle(f"{mapper_path}/icd10.pkl")
