==> src/01_claim_data_cleaning.ipynb <==
# Cleaning Claim Data
## 01_claim_data_cleaning

| Date | User | Change Type | Remarks |  
| ---- | ---- | ----------- | ------- |
| 18/09/2025   | Adrienne | Created   | Created to flatten data | 
| 27/09/2025 | Adrienne | Updated | Added Martin's code and other updates |
| 13/10/2025 | Martin | Updated | Code cleanup. Added docstrings for functions |

## Content

* [Preprocess JSON](#preprocess-JSON)
  * [Functions to Flatten JSON File](#functions-to-Flatten-JSON-File)
  * [Specific column processing](#specific-column-processing)
* [Drop Columns](#drop-columns)

## Claims

__Columns__

- `contained` - birthDate, extension, gender, id, identifier, name, resourceType, id, identifier, resourceType
- `created`
- `diagnosis` - [diagnosisCodeableConcept, sequence, type] x 23
- `extension`
- `id`
- `resourceType`
- `status`
- `supportingInfo`
- `type`
- `use`
- `billablePeriod_end`
- `billablePeriod_start`
- `facility_extension`
- `identifier_system`
- `identifier_type`
- `identifier_value`
- `insurance_coverage`
- `insurance_focal`
- `insurance_sequence`
- `item_extension`
- `item_productOrService`
- `item_revenue`
- `item_sequence`
- `item_servicedDate`
- `meta_lastUpdated`
- `patient_reference`
- `priority_coding`
- `provider_reference`
- `total_currency`
- `total_value`
- `contained_identifer_patient_medicare_number`
- `contained_name_family`
- `contained_name_given`

import pandas as pd
import seaborn as sns
import matplotlib.pyplot as plt
from io import StringIO
import os
import json
from collections import OrderedDict
import pickle
import numpy as np

# Load JSON files
path = "../data/raw"
# claim = pd.read_json(f"{path}/Claim.ndjson", lines=True)
claim = pd.read_json(f"{path}/Claim.ndjson", lines=True)

## Preprocess JSON

The claims file has complex nesting. The code below prints out a single row of data so the structure can be easily seen.

Alternatively, a single entry can be explored by pasting the JSON file into a viewer like [here](https://jsonformatter.org/json-viewer)

claim_text = claim.head(1)
# claim_text = claim.head(93532).tail(1)
for key, value in claim_text.items():
    print(f"key*: {key}")
    for item in value:
            if isinstance(item, dict):
                for key, value in value.items():
                    print(f"\tkey:: ({key}) value: ({value})\n")
            elif isinstance(item, list):
                for i in item:
                    if isinstance(i, dict):
                        for key, value in i.items():
                            if isinstance(value, list):
                                for y in value:
                                    print(f"\t\tkey: {key} list: ({y})\n")
                            else:
                                print(f"\t\tkey: ({key}) value: ({value})\n")
                    else:
                        print(f"{i}\n")
            else:
                print(f"\tvalue$: {item}\n")


### Functions to Flatten JSON File

def flatten_json(nested_json: dict, prefix: str='') -> OrderedDict:
    """
    Recursively flattens a nested JSON object or dictionary into a single level.

    Notes:
        - Nested dictionaries and lists are flattened such that keys from deeper levels
          in the hierarchy are concatenated with underscores
        - Lists of dictionaries are handled by appending index numbers to the keys.
        - Non-dict lists are serialized using JSON encoding
        - Returns OrderedDict, a flattened version of the input json, where keys represent
          the nested structure and values are the corresponding data

    Args:
        nested_json (dict): JSON object
        prefix (str, optional): Prefix for keys. Defaults to ''.

    Returns:
        OrderedDict: unnested JSON object
    """
    out = OrderedDict()
    for key, value in nested_json.items():
        if isinstance(value, dict):
            # Recursively flatten nested dictionaries
            out.update(flatten_json(value, prefix + key + '_'))
        elif isinstance(value, list):
            if len(value) > 0:
                if isinstance(value[0], dict):
                    # Handle list of dictionaries by flattening each item
                    for i, item in enumerate(value):
                        out.update(flatten_json(item, prefix + key + '_' + str(i) + '_'))
                else:
                    # Non-dict lists are serialized into a JSON string
                    out[prefix + key] = json.dumps(value)
            else:
                # Empty lists are serialized as JSON strings
                out[prefix + key] = json.dumps(value)
        else:
            # Base case: key-value pair where value is not a list or dict
            out[prefix + key] = value
    return out

def process_dataframe(df: pd.DataFrame) -> pd.DataFrame:
    """
    Processes a pandas DataFrame by flattening any JSON-like data (dictionaries or lists)
    present in its columns and converting it into a new DataFrame.

    Notes:
        - The function iterates through each row and flattens any JSON-like data (dictionaries or lists)
        - Non-nested data is left unchanged
        - The resulting DataFrame will contain a combination of original columns and
          additional columns derived from the flattened structure
        - Returns a new pandas DataFrame with the flattened data

    Args:
        df (pd.DataFrame): Dataframe containing JSON objects in columns

    Returns:
        pd.DataFrame: Flattened dataframe
    """
    flattened_data = []
    for _, row in df.iterrows():
        flattened_row = {}
        for column, value in row.items():
            if isinstance(value, (dict, list)):
                # Flatten any dictionary or list
                flattened = flatten_json({column: value})
                flattened_row.update(flattened)
            else:
                # Keep non-nested columns unchanged
                flattened_row[column] = value
        flattened_data.append(flattened_row)
    return pd.DataFrame(flattened_data)


# Flatten claim_df
flat_claim_df = process_dataframe(claim)

### Specific Column Processing

Additional processing required for specific columns. Functions involve cleaning up text data and renaming

# Preprocess other columns
flat_claim_df['patient_medicare_number'] = flat_claim_df['contained_0_identifier_0_value']
flat_claim_df['patient_first_name'] = flat_claim_df['contained_0_name_0_given'].str.replace(r'[ \[ \]"]', '', regex=True)
flat_claim_df['patient_last_name'] = flat_claim_df['contained_0_name_0_family']
flat_claim_df['unique_claim_ID'] = flat_claim_df['identifier_0_value'].str.replace(r'[-]', '', regex=True).str.replace('dcn', '')
flat_claim_df['drg_code'] = flat_claim_df['supportingInfo_1_code_coding_0_code']
flat_claim_df['provider_number'] = flat_claim_df['contained_1_identifier_0_value']
flat_claim_df['national_provider_identifier'] = flat_claim_df['contained_1_identifier_1_value']
flat_claim_df['type_of_bill'] = flat_claim_df['supportingInfo_0_code_coding_0_code']
flat_claim_df['claim_type'] = flat_claim_df['type_coding_0_code']
flat_claim_df['location_of_bill'] = flat_claim_df['facility_extension_0_valueCoding_code']
flat_claim_df['gender'] = flat_claim_df['contained_0_gender']

diagnosis_cols = [col for col in flat_claim_df.columns if 'diagnosisCodeableConcept_coding_0_code' in col]

# Create list column of diagnoses
flat_claim_df['diagnosis_ls'] = flat_claim_df[diagnosis_cols].apply(lambda row: [x for x in row if pd.notnull(x)] , axis = 1)

# Quick check
diagnosis_cols.append('diagnosis_ls')
flat_claim_df[diagnosis_cols].head(3)

# Sometimes there are admitting and/or principle diagnosis code types.  will make separate columns for them
diagnosis_code_type_cols = flat_claim_df.columns[flat_claim_df.columns.str.contains('type_0_coding_0_code|type_1_coding_0_code')]

for index, row in flat_claim_df.iterrows():
    for col in diagnosis_code_type_cols:       
        if row[col] == 'admitting':
            ind = col[11:12]
            if ind == '_':
                ind = col[0:12]
            else:
                ind = col[0:13]
            col_name = ind + 'diagnosisCodeableConcept_coding_0_code'
            flat_claim_df.loc[index, 'admitting_diagnosis'] = row[col_name]
        elif row[col] == 'principal':
            ind = col[11:12]
            if ind == '_':
                ind = col[0:12]
            else:
                ind = col[0:13]
            col_name = ind + 'diagnosisCodeableConcept_coding_0_code'
            flat_claim_df.loc[index,'principal_diagnosis'] = row[col_name]                  

flat_claim_df.head()

# Create list column of HCPCS codes
hcpcs_cols = [col for col in flat_claim_df.columns if 'productOrService_coding_0_code' in col]

flat_claim_df['hcpcs_ls'] = flat_claim_df[hcpcs_cols].apply(lambda row: [x for x in row if pd.notnull(x)] , axis = 1)

# Quick check
hcpcs_cols.append('hcpcs_ls')
flat_claim_df[hcpcs_cols].head(3)

procedure_cols = [col for col in flat_claim_df.columns if 'procedureCodeableConcept_coding_0_code' in col]
#procedure_cols

# Create list column of diagnoses
flat_claim_df['procedure_ls'] = flat_claim_df[procedure_cols].apply(lambda row: [x for x in row if pd.notnull(x)] , axis = 1)

# Quick check
procedure_cols.append('procedure_ls')
flat_claim_df[procedure_cols].head(8500).tail(5)

### Drop Columns

bef_len = str(len(flat_claim_df.columns))

# Remove all columns that have system or url in the name
drop_cols = flat_claim_df.columns[flat_claim_df.columns.str.contains("system|url")]
flat_claim_df = flat_claim_df.drop(drop_cols, axis = 1)

# Remove all columns that have extension in the name
ext_cols = sorted([col for col in flat_claim_df.columns if 'extension' in col])
flat_claim_df = flat_claim_df.drop(ext_cols, axis = 1)

print(f'number of columns before dropping: {bef_len}')
print(f'number of columns after dropping: {len(flat_claim_df.columns)}')

flat_claim_df.head()

# Creating claim sample file and saving all files to pickle
path = "../data/clean"
flat_claim_df.to_pickle(f"{path}/claim.pkl")
