==> src/02_EDA_patient.ipynb <==
# EDA of Patient Data
## 02_EDA_Patient

| Date | User | Change Type | Remarks |  
| ---- | ---- | ----------- | ------- |
| 28/09/2025 | Adrienne | Created   | Moved Martin's code to this program | 
| 13/10/2025 | Martin | Updated   | Code cleanup | 

## Content

* [Load Data](#load-data)
* [Patient Data Exploration](#patient-data-exploration)

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import re
from datetime import datetime
import json_lines
import seaborn as sns

### Load Data

path = "../data/clean"
claim = pd.read_pickle(f"{path}/claim.pkl")
patient = pd.read_pickle(f"{path}/patient.pkl")
eob = pd.read_pickle(f"{path}/eob.pkl")
# coverage = pd.read_pickle(f"{path}/coverage.pkl")

### Patient Data Exploration

What does a single patient "look" like across all our data?

# Select relevant columns for initial exploration
patient = patient[['address_0_state', 'birthDate', 'deceasedBoolean', 'extension_1_valueCoding_display', 'extension_3_valueDate', 'gender', 'id', 'identifier_1_value', 'deceasedBoolean', 'deceasedDateTime', 'address_0_postalCode']]
patient = patient.rename({
  'extension_1_valueCoding_display': 'race',
  'extension_3_valueDate': 'referenceYear',
  'identifier_1_value': 'patient_medicare_number',
  'address_0_postal_code': 'postalCode'
}, axis=1)

patient.head()

patient_id = -10000000000322
medicare_number = "1S00E00AA50"

c = claim[claim['patient_medicare_number'] == medicare_number]
e = eob[eob['id'] == str(patient_id)]
p = patient[patient['patient_medicare_number'] == medicare_number]

c.unique_claim_ID.head()

# Checking if there are any non-na DRGs with HCPCS codes
c['hcpcs_0'] = c['hcpcs_ls'].str[0]
c[(~c['drg_code'].isna()) & (~c['hcpcs_0'].isna())]

#c[c['unique_claim_ID'] == '-100007150']
c[c['unique_claim_ID'] == '100005174']

# print(f"Patient - Medicare Number: {medicare_number}")
# print(f"Race: {p['race']}")
# print(f"Gender: {p['gender']}")
# print(f"Birth Date: {p['birthDate']}")
# print(f"Deceased: {p['deceasedBoolean'] == 1}")
# print(f"Postal Code: {p['address_0_postalCode']}")
